Task #4959: Reintegrated task branch (DelayAndBandPassKernel: wrong subbandFreq type passed) with the trunk

+++ RTCP/Cobalt/GPUProc/src/cuda/Kernels/DelayAndBandPassKernel.h
@@ -88,7 +88,7 @@ namespace LOFAR
 
 
       void enqueue(PerformanceCounter &counter,
-                   float subbandFrequency, size_t SAP);
+                   double subbandFrequency, size_t SAP);
     };
 
     // Specialization of the KernelFactory for
